requirements.txt
bottleneck==1.4.2
brotli==1.0.9
pyqt6==6.7.1
pyqt6-sip==13.9.1
pyyaml==6.0.2
asttokens==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.11.0
debugpy==1.8.14
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
fonttools==4.55.3
importlib-metadata==8.7.0
ipykernel==6.29.5
ipython==8.37.0
jedi==0.19.2
jupyter-client==8.6.3
jupyter-core==5.8.1
kiwisolver==1.4.8
llvmlite==0.44.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
mkl-fft==1.3.11
mkl-random==1.2.8
mkl-service==2.4.0
nest-asyncio==1.6.0
numba==0.61.0
numexpr==2.10.1
numpy==2.1.3
opencv-python==4.10.0
opencv-python-headless==4.10.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pickleshare==0.7.5
pillow==11.1.0
pip==25.1
platformdirs==4.3.8
prompt-toolkit==3.0.51
psutil==7.0.0
pure-eval==0.2.3
pygments==2.19.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==307
pyzmq==26.2.0
seaborn==0.13.2
setuptools==72.1.0
shapely==2.0.6
sip==6.10.0
six==1.17.0
stack-data==0.6.3
tomli==2.0.1
tornado==6.4.2
traitlets==5.14.3
typing-extensions==4.14.0
tzdata==2025.2
unicodedata2==15.1.0
wcwidth==0.2.13
wheel==0.45.1
zipp==3.22.0
zmq==0.0.0
autocommand==2.2.2
backports.tarfile==1.2.0
importlib-resources==6.4.0
inflect==7.3.1
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
more-itertools==10.3.0
ordered-set==4.1.0
typeguard==4.3.0
